Updated aws and k8s modules
