+++ modules/aws/src/main/scala/etlflow/task/S3SensorTask.scala
@@ -7,11 +7,11 @@ import zio.{Clock, RIO, ZIO}
 import scala.concurrent.duration._
 
 case class S3SensorTask(name: String, bucket: String, key: String, retry: Int, spaced: Duration)
-    extends EtlTask[S3Env with Clock, Unit] {
-  override protected def process: RIO[S3Env with Clock, Unit] = for {
+    extends EtlTask[S3 with Clock, Unit] {
+  override protected def process: RIO[S3 with Clock, Unit] = for {
     _ <- ZIO.logInfo("#" * 50)
     _ <- ZIO.logInfo(s"Starting sensor for s3 location s3://$bucket/$key")
-    _ <- S3Api
+    _ <- S3
       .lookupObject(bucket, key)
       .flatMap { out =>
         if (out) ZIO.logInfo(s"Found key $key in s3 bucket s3://$bucket")
