+++ modules/k8s/src/main/scala/etlflow/k8s/K8S.scala
@@ -7,10 +7,9 @@ import com.coralogix.zio.k8s.model.batch.v1.{Job, JobSpec}
 import com.coralogix.zio.k8s.model.pkg.apis.meta.v1.ObjectMeta
 import zio.{RIO, TaskLayer, ULayer, ZIO, ZLayer}
 
-@SuppressWarnings(Array("org.wartremover.warts.ToString"))
-case class K8S() extends K8SApi.Service {
+trait K8S {
 
-  /** Method: createJob - Submit a job to kubernetes cluster using given job configurations
+  /** Method: createJob - Submit a job to kubernetes cluster using given job configuration
     * @param metadata
     *   kubernetes job metadata
     * @param spec
@@ -20,10 +19,7 @@ case class K8S() extends K8SApi.Service {
     * @return
     *   Job
     */
-  def createJob(metadata: ObjectMeta, spec: JobSpec, namespace: K8sNamespace): RIO[Jobs, Job] =
-    ZIO
-      .environmentWithZIO[Jobs](_.get.create(Job(metadata = Some(metadata), spec = Some(spec)), namespace))
-      .mapError(e => new Exception(e.toString))
+  def createJob(metadata: ObjectMeta, spec: JobSpec, namespace: K8sNamespace): RIO[Jobs, Job]
 
   /** Method: getJob - Get a kubernetes job details for given job name
     * @param name
@@ -33,14 +29,35 @@ case class K8S() extends K8SApi.Service {
     * @return
     *   Job
     */
-  def getJob(name: String, namespace: K8sNamespace): RIO[Jobs, Job] =
-    ZIO
-      .environmentWithZIO[Jobs](_.get.get(name, namespace))
-      .mapError(e => new Exception(e.toString))
-
+  def getJob(name: String, namespace: K8sNamespace): RIO[Jobs, Job]
 }
-
 object K8S {
-  val live: TaskLayer[K8sEnv] = (k8sDefault >>> Jobs.live) ++ ZLayer.succeed(K8S())
-  val test: ULayer[K8sEnv]    = Jobs.test ++ ZLayer.succeed(K8S())
+
+  /** Method: createJob - Submit a job to kubernetes cluster using given job configuration
+    * @param metadata
+    *   kubernetes job metadata
+    * @param spec
+    *   kubernetes job spec
+    * @param namespace
+    *   kubernetes cluster namespace, defaults to default namespace
+    * @return
+    *   Job
+    */
+  def createJob(metadata: ObjectMeta, spec: JobSpec, namespace: K8sNamespace = K8sNamespace.default): RIO[K8S with Jobs, Job] =
+    ZIO.environmentWithZIO[K8S](_.get.createJob(metadata, spec, namespace))
+
+  /** Method: getJob - Get a kubernetes job details for given job name
+    * @param name
+    *   kubernetes job name
+    * @param namespace
+    *   kubernetes cluster namespace, defaults to default namespace
+    * @return
+    *   Job
+    */
+  def getJob(name: String, namespace: K8sNamespace = K8sNamespace.default): RIO[K8S with Jobs, Job] =
+    ZIO.environmentWithZIO[K8S](_.get.getJob(name, namespace))
+
+  val live: TaskLayer[K8S with Jobs] = (k8sDefault >>> Jobs.live) ++ ZLayer.succeed(K8SImpl())
+
+  val test: ULayer[K8S with Jobs] = Jobs.test ++ ZLayer.succeed(K8SImpl())
 }
