--- modules/k8s/src/main/scala/etlflow/k8s/package.scala
@@ -1,7 +1,0 @@
-package etlflow
-
-import com.coralogix.zio.k8s.client.batch.v1.jobs.Jobs
-
-package object k8s {
-  type K8sEnv = K8SApi.Service with Jobs
-}
